Fix player movement

+++ include/Systems/PhysicsSystem.h
@@ -11,7 +11,7 @@ class PhysicsSystem : public System {
 private:
     void _update(int dt) override;
 
-    static void updatePlayer(PlayerComponent& player, TransformationComponent& transform, VelocityComponent& velocity, int dt);
+    static void updatePlayer(PlayerComponent& player, TransformationComponent& transform, VelocityComponent& velocity, float dt);
 
 public:
     PhysicsSystem(entt::registry* registry);
